Proctor Data Analysis Assessment
1 Download and process the data
Suggested time: 15 minutes

You can download a clean version of two related datasets from this study. Go to the following link, and click on “Download dataset” on the right hand side of the page.

https://datadryad.org/stash/dataset/doi:10.7272/Q6DZ06J3

It will download a .zip file that includes the source code and data from a secondary analysis of this trial. The datasets are in both comma-separated values (.csv) and R dataset (.rds) formats. The archive includes an associated codebook for each dataset.

mbita_schisto.csv/.rds: child level data

mbita_spatial.csv/.rds: additional community-level characteristics

Review the data and codebooks. Familiarize yourself with them. Then, read in the data. (These data are very clean so unlike most projects there is not an extensive amount of data processing required. In reality, data processing is often the most time consuming part of a project!) Depending on how you approach the work below, if you want to join the two datasets the 1:many key variable is village id (vid).

Create a derived variable that is an indicator of whether the child was seropositive to either the SEA antigen or the Sm25 antigen. We will be using a combined measure of serpositivity as the outcome!

2 Describe the data
Suggested time: 30-45 minutes

A key step in any data analysis is getting familiar with the dataset. Some of that process can involve simply looking at the data itself (e.g., in RStudio, View() can be very helpful in this regard!).

Provide some simple descriptive summaries that help describe the data. For example, how many children were measured at baseline (2012) and in subsequent survey visits based on blood- and stool-based measures of infection? How complete were the measurements? For quantitative variables, is there anything you see about their distribution that might be important to consider in any downstream analysis?

There is no one way to do this and we are not looking for a specific result. We are interested to learn how you would describe the data to have a sense for what it contains.

3 Summarize baseline characteristics
Suggested time: 45-60 minutes

In randomized, controlled trials (RCTs) we use random allocation of treatment to balance measurable and unmeasurable characteristics between treatment groups. On average, the potential outcomes in the two groups should be the same in the absence of treatment — any differences we observe in outcomes can be attributed to a treatment effect.

One important step in an RCT is to compare groups based on measurable baseline characteristics. This is often “Table 1” in reporting for RCTs and is item 15 on the CONSORT checklist for cluster randomized trials: https://www.bmj.com/content/345/bmj.e5661 .

The schistosomiasis study was a community randomized trial so the independent units for analysis are the community.

Create a table that summarizes individual-level and cluster-level characteristics by randomized group (community-wide treatment and school-based treatment). Each row should be a variable or level of that variable. There should be a separate column for each group. For measures of S. mansoni infection or antibody response, limit your summary to the categorical measures rather than quantitative measures. For other quantitative variables, summarize the mean and standard deviation (and/or median and interquartile range if you feel that is more appropriate). For categorical variables, report the N and percent.

Do the groups look well balanced at baseline based on measured characteristics? Briefly explain why you think they are or are not well balanced.

Idea for reflection: what element of the design could contribute to better or worse balance between groups in their baseline characteristics?